revert to old pipliene but use testing nomenclature in test folder
