+++ .github/workflows/test.yml
@@ -165,19 +165,17 @@ jobs:
 
       - name: Print Frontend Coverage Report
         run: cat frontend/coverage/lcov.info || echo "Frontend coverage report not found"
-       
+
       - name: Run SonarCloud Analysis
         run: |
           export PATH="/opt/sonar-scanner/bin:$PATH"
           sonar-scanner \
             -Dsonar.organization=${{ secrets.SONAR_ORG }} \
             -Dsonar.projectKey=Andrew-Dagher_SOEN-390 \
+            -Dsonar.sources=. \
             -Dsonar.host.url=${{ secrets.SONAR_HOST_URL }} \
             -Dsonar.token=${{ secrets.SONAR_TOKEN }} \
-            -Dsonar.sources=backend,frontend/app \
-            -Dsonar.exclusions=**/__tests__/** \
-            -Dsonar.coverage.exclusions=**/__tests__/** \
-            -Dsonar.javascript.lcov.reportPaths=backend/coverage/lcov.info,frontend/coverage/lcov.info
+            -Dsonar.javascript.lcov.reportPaths=backend/coverage/lcov.info,frontend/coverage/lcov.info || (echo "SonarCloud analysis failed" && exit 1)
 
       - name: Get SonarCloud Results
         run: "echo 'Check SonarCloud Report: https://sonarcloud.io/dashboard?id=Andrew-Dagher_SOEN-390&branch=${GITHUB_REF##*/}'"
