+++ frontend/__tests__/BottomNavBar.test.jsx
